QuickFit3: several parts are more stable now, if large images are processed (>65535 pixels) QuickFit3/imFCS: added two functions (to data menu in fit parameter image RDR) that allow to copy the average CorrelationFunction and the selected pixels correlation function to a table RDR
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3128 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ lib/qfimporterimageseries.cpp
@@ -14,7 +14,7 @@ QFImporterImageSeries::QFImporterImageSeries() {
     averageBinning=false;
 }
 
-uint16_t QFImporterImageSeries::frameWidth() {
+uint32_t QFImporterImageSeries::frameWidth() {
     if (binning<1) binning=1;
     int xx0=x0;
     int xx1=x1;
@@ -36,7 +36,7 @@ uint16_t QFImporterImageSeries::frameWidth() {
     return w;
 }
 
-uint16_t QFImporterImageSeries::frameHeight() {
+uint32_t QFImporterImageSeries::frameHeight() {
     if (binning<1) binning=1;
     int yy0=y0;
     int yy1=y1;
@@ -95,8 +95,8 @@ bool QFImporterImageSeries::readFrameFloat(float* data) {
     if (yy1>intFrameHeight()-1) yy1=intFrameHeight()-1;
     if (yy0>yy1) qSwap(yy0, yy1);
 
-    uint16_t w=frameWidth();
-    uint16_t h=frameHeight();
+    uint32_t w=frameWidth();
+    uint32_t h=frameHeight();
     int hh=abs(yy1-yy0)+1;
     int ww=abs(xx1-xx0)+1;
 
@@ -172,8 +172,8 @@ bool QFImporterImageSeries::readFrameDouble(double *data) {
     if (yy1>intFrameHeight()-1) yy1=intFrameHeight()-1;
     if (yy0>yy1) qSwap(yy0, yy1);
 
-    uint16_t w=frameWidth();
-    uint16_t h=frameHeight();
+    uint32_t w=frameWidth();
+    uint32_t h=frameHeight();
     int hh=abs(yy1-yy0)+1;
     int ww=abs(xx1-xx0)+1;
 
@@ -249,8 +249,8 @@ bool QFImporterImageSeries::readFrameUINT16(uint16_t* data) {
     if (yy1>intFrameHeight()-1) yy1=intFrameHeight()-1;
     if (yy0>yy1) qSwap(yy0, yy1);
 
-    uint16_t w=frameWidth();
-    uint16_t h=frameHeight();
+    uint32_t w=frameWidth();
+    uint32_t h=frameHeight();
     int hh=abs(yy1-yy0)+1;
     int ww=abs(xx1-xx0)+1;
 
@@ -261,7 +261,7 @@ bool QFImporterImageSeries::readFrameUINT16(uint16_t* data) {
     }
 
     for (register int i=0; i<w*h; i++) data[i]=0;
-    uint16_t binAvg=1.0;
+    uint32_t binAvg=1.0;
     if (averageBinning) binAvg=binning*binning;
 
     if (!interleavedBinning) {
