+++ plugins/basicimageimporters/qfimagereaderdiffusion4binary.h
@@ -45,9 +45,9 @@ class QFImageReaderDiffusion4Binary: public QFImporterImageSeries
 
     protected:
         /** \brief return the width of the frames (valid after open() returned \c true */
-        virtual uint16_t intFrameWidth();
+        virtual uint32_t intFrameWidth();
         /** \brief return the height of the frames (valid after open() returned \c true */
-        virtual uint16_t intFrameHeight();
+        virtual uint32_t intFrameHeight();
         /** \brief read a new frame into the given array of floating point numbers */
         virtual bool intReadFrameFloat(float* data);
         /** \brief read a new frame into the given array of integers */
@@ -55,8 +55,8 @@ class QFImageReaderDiffusion4Binary: public QFImporterImageSeries
         virtual bool intReadFrameDouble(double* data);
         QFile input;
         QString filename;
-        uint16_t width;
-        uint16_t height;
+        uint32_t width;
+        uint32_t height;
         uint64_t frame;
         uint64_t frames;
         double timestep;
