+++ plugins/basicimageimporters/qfimagereadertinytiff.h
@@ -47,9 +47,9 @@ class QFImageReaderTinyTIFF: public QFImporterImageSeries
 
     protected:
         /** \brief return the width of the frames (valid after open() returned \c true */
-        virtual uint16_t intFrameWidth();
+        virtual uint32_t intFrameWidth();
         /** \brief return the height of the frames (valid after open() returned \c true */
-        virtual uint16_t intFrameHeight();
+        virtual uint32_t intFrameHeight();
         /** \brief read a new frame into the given array of floating point numbers */
         virtual bool intReadFrameFloat(float* data);
         /** \brief read a new frame into the given array of integers */
@@ -57,8 +57,8 @@ class QFImageReaderTinyTIFF: public QFImporterImageSeries
         virtual bool intReadFrameDouble(double* data);
         TinyTIFFReaderFile* tif;
         QString filename;
-        uint16_t width;
-        uint16_t height;
+        uint32_t width;
+        uint32_t height;
         uint32_t frame;
 
     public:
