+++ plugins/interfaces/qfrdroverviewimageinterface.h
@@ -59,12 +59,12 @@ class QFRDROverviewImageInterface {
         /** \brief returns the name of the given image */
         virtual QString getOverviewImageName(int image) const=0;
 
-        /*! \brief return a pointer to a uint16_t image of size getDataImageWidth() * getDataImageHeight()
+        /*! \brief return a pointer to a double image of size getDataImageWidth() * getDataImageHeight()
 
             This image can be used to allow the user to select points in the image, or just as an overview. E.g. for
             imFCS-data this could simply be the averaged intensity of every pixel.
 
-            \return  a uint16_t image of size getDataImageWidth() * getDataImageHeight(), or NULL if no image is available
+            \return  a double image of size getDataImageWidth() * getDataImageHeight(), or NULL if no image is available
          */
         virtual double* getOverviewImage(int image) const=0;
         /*! \brief returns a list of geometrix elements overlayn to the image */
